Ya casi está resuelto todo el asunto de las estadisticas, solo faltaría hacer bien lo de exportar el .csv

+++ servicioEstadistica/src/main/java/ar/edu/utn/frba/dds/model/dtos/DatoEstadisticoDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class DatoEstadisticoDTO {
   private String primerCriterio;
   private String segundoCriterio;
-  private Long hora;
+  private Integer hora;
   private Long cantidad;
   private Double porcentaje;
   private Long total;
